[ui] Check engine is alive before running teardown queries (e.g. drop table/view) in Track.onDestroy() hooks.
Bug: 315096958
(cherry picked from commit 2fe5cb2726f51cadd2f1d260da7a9c8684a50edf)

+++ ui/src/tracks/custom_sql_table_slices/index.ts
@@ -92,7 +92,9 @@ export abstract class CustomSqlTableSliceTrack<
         });
     await this.engine.query(sql);
     return DisposableCallback.from(() => {
-      this.engine.query(`DROP VIEW ${this.tableName}`);
+      if (this.engine.isAlive) {
+        this.engine.query(`DROP VIEW ${this.tableName}`);
+      }
     });
   }
